Change Sprite and Display to use uint16_t for dimensions

+++ src/Display/Display.cpp
@@ -1,6 +1,6 @@
 #include "Display.h"
 
-Display::Display(uint8_t width, uint8_t height, int8_t blPin, int8_t rotation, bool mirror) : tft(), blPin(blPin),
+Display::Display(uint16_t width, uint16_t height, int8_t blPin, int8_t rotation, bool mirror) : tft(), blPin(blPin),
 		width(width), height(height), baseSprite(new Sprite(&tft, width, height)), rotation(rotation), mirror(mirror){
 
 }
